### 2021-01-19 #2972 @whedon
[PRE REVIEW]: Unsupervised learning approach towards anomaly detection in compat logs with ADE
**Submitting author:** @ayush-1506 (<a href="http://orcid.org/0000-0003-3550-691X">Ayush Shridhar</a>)
**Repository:** <a href="https://github.com/openmainframeproject/ade.git" target ="_blank">https://github.com/openmainframeproject/ade.git</a>
**Version:** v1.0.5
