### 2021-01-20 #2972 @ayush-1506
@kthyng Yes, the code and paper live inside a different branch of the repository.
Link : https://github.com/openmainframeproject/ade/tree/logs

Can we get whedon to use this branch instead of master? Else I'll discuss with my collaborators to merge this into master as soon as possible.
