### 2021-01-20 #2972 @kthyng
@ayush-1506 Yes it is fine to have the paper in another branch. Please look through the [paper requirements](https://joss.readthedocs.io/en/latest/submitting.html#) to be sure you've covered them all. For one thing, we require a section entitled "Statement of Need".
