### 2021-01-20 #2972 @kthyng
@ayush-1506 This looks like interesting work, but can you make a compelling argument for why it is research software in particular? You can read more about that [requirement here](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). I'm going to label this with a scope query to get the editorial board's input on this, which should take 1-2 weeks.
