### 2021-01-20 #2972 @ayush-1506
@kthyng Thanks for the input. I'll add a _Statement of Need_ section (which will support the fact that this software and approach solves a problem). Do I need to add the argument behind this being a research software in the paper or a comment here will suffice?
