### 2021-01-21 #2972 @ayush-1506
Made changes to the paper, summarizing the same here:

### Objective:
The aim of the project is to solve the problem of efficiently detecting anomalous logs slices from large set of logs (This can include sparse logs such as Linux Syslogs RFC3164/RFC5424 format or very dense logs such as those generated from Spark jobs). We introduce a novel approach towards solving this problem. Expanding on the approach later in this comment.

### Need:
Why do we need to find anomalous log slices?

Debugging system failures is a cumbersome task. Upheaval behavior in the system can be identified by studying the logs generated while the system was running. If the system fails or reacts with unexpected behavior, this data is logged somewhere. However, going through hours of dense logs is a challenge: sysadmins typically need to race against time to study large amounts of  log messages to decipher the root cause of the issue. Such system failures are very common and at times unavoidable. Over the years, such behavior and crashes have led to huge loss of time and resources.

### Relevant work and our approach:

While there has been work towards this direction of anomaly detection in large logs, such as TadGAN (https://arxiv.org/abs/2009.07769) and semi-supervised adversarial learning with GANs (https://doi.org/10.1109/ciss.2019.8693024), most of these approaches have focused on using large deep learning models and some treat this as a supervised problem. These models are large to train and also comparatively slower. On the other hand, we treat the problem as a statistical one and use unsupervised learning techniques for fast and robust detection of anomalous slices. Being an unsupervised approach, we don't need labelled features. Avoiding computationally heavy deep learning makes our system fast and it's written in the Java language which makes it ideal for enterprise IT use cases (which can be adapted to others too).  To this end, we divide the problem into 3 main categories:

- Unsupervised learning algorithms: At the heart of ADE are unsupervised learning algorithms that are trained to understand the actual expected behavior of the system and compare it with the observed behavior during inference.

- Model groups: We divide the training into several categories, called _model groups_.

- Statistical scores: To come up with an anomaly score, we calculate a number of statistical score that contribute to the final anomaly score. These scores include _Interval Anomaly Score, Bernoulli Score, Poisson Score, LogNormal Score, Best-of-two score, rarity score, severity score, Clustering score, Percentile score, FullBernoulliClusterAware score_ to name a few.

Kindly let me know if I messed something or if there are any questions here.

Along with this, for each message, we try to classify it into four categories based on the frequency of the particular family of messages. These classes include:

- New : Defines a completely new message (previously unseen)
- IN_SYNC : Implies that ADE expects the message to be issued in a periodic pattern and the message was issued as expected
- NOT_IN_SYNC : Implies that ADE expects the message to be issued in a periodic pattern but the message was not expected
- NOT_PERIODIC : Indicates that ADE does not expect the message to be periodic

Using all this calculated information, we allocate an anomaly score to every internal slice. The higher the anomaly score, the greater are the chances of that particular slice being the source of anomalous logs.

### Output format:

Output format: Finally, we write out the analysis output in XML format. An example of the analysis output for a day can be seen [here](https://openmainframeproject.github.io/ade/20151212/index.xml). We also provide specialized output for each interval, which can be accessed by clicking on the XML links associated with each slice. Examples of analysis for a period can be viewed [here](https://openmainframeproject.github.io/ade/20151212/intervals/interval_0.xml). Our approach has shown comparatively accurate results when tested on real data, along with fast inference and training.
