### 2021-01-24 #2972 @danielskatz
@ayush-1506  - can you explain what code you are submitting to JOSS in this branch vs the overall repo?  The paper seems to describe ADE, which is what the repo contains, but you also have suggested that https://github.com/openmainframeproject/ade/tree/logs is the contribution being submitted here, and I can't tell if the paper describes that specific contribution.
