### 2021-01-26 #2972 @ayush-1506
@danielskatz There were some issues ([here](https://github.com/openmainframeproject/ade/pull/70#issuecomment-760617398) : CLA wasn't registering the contributors) with pushing new commits to master branch in the ADE repository, hence all development was being done and reviewed in the _logs_ branch temporarily. However, the issues with CLA have been resolved now and all changes have been merged to master. We can take the master branch as the main branch with paper and code from now on.
