### 2021-01-26 #2972 @danielskatz
So is all of the content in the main branch the JOSS submission?
