### 2021-01-26 #2972 @ayush-1506
> So is all of the content in the main branch the JOSS submission?

@danielskatz Yes: all changes were merged yesterday.
