### 2021-02-18 #2972 @arcuri82
Hi,

@gkthiruvathukal : I made a first review at https://github.com/openmainframeproject/ade/issues/85  However, as I do not use Linux, I did not run the software (just read documentation, compiled software, run its test cases). We would need the second reviewer to make sure to run the software. In worst case, I can try to install a virtual machine to run Linux on it (but that would be quite a bit of work, so might take me a while...). However, the authors have quite an extensive documentation, with few examples of outputs
