### 2021-03-01 openmainframeproject/ade#86 @mdpiper
JOSS review: Editing pass on paper
This issue is raised in reference to JOSS submission [3052](https://github.com/openjournals/joss-reviews/issues/3052).

There are numerous spelling errors, small typos, and omissions in the submitted paper. It would benefit from another editing pass by the authors.
