### 2021-03-01 openmainframeproject/ade#86 @mdpiper
A pair of suggestions that may help address this issue:

1. Run the text through `ispell`
2. For a quick grammar check, paste the text into Google Docs or Word

(I realize that the authors are very likely aware of these solutions, but I just want to be careful and thorough in my review.)
