### 2021-03-02 openmainframeproject/ade#87 @mdpiper
JOSS review: Writing style in paper
This issue is raised in reference to JOSS submission [3052](https://github.com/openjournals/joss-reviews/issues/3052).

The writing style in the paper is colloquial; in particular, the sections _Statement of Need_ and _Novelties_. The language is understandable (I enjoyed reading the paper!), but it's suited more for a magazine article or a blog post. Since JOSS is an academic journal, the language needs to be more precise. The paper in general, and the sections mentioned above in particular, should be edited to reduce verbosity and use more precise language.
