### 2021-03-02 openmainframeproject/ade#87 @mdpiper
A suggestion that may help: take a look at some recent JOSS papers to see the tone that's used.
