### 2021-03-02 openmainframeproject/ade#88 @mdpiper
JOSS review: State of the field
This issue is raised in reference to JOSS submission [3052](https://github.com/openjournals/joss-reviews/issues/3052).

The paper includes a paragraph on others' work on automated log anomaly detection. However, the cited articles are a conference paper and a pair of arXiv preprints. Are there any refereed articles on this topic that can be cited?
