### 2021-03-08 openmainframeproject/ade#89 @mdpiper
JOSS review: Performance
This issue is raised in reference to JOSS submission [3052](https://github.com/openjournals/joss-reviews/issues/3052).

There's an unsubstantiated performance claim on line 48 of the manuscript. The claim should be backed up with data or struck.
