### 2021-03-08 openmainframeproject/ade#89 @mdpiper
A suggestion: A quick comparison between ADE and the other techniques might be all that's necessary. E.g.,

> We ran ADE, along with approach A and B, over N time slices of a dataset. ADE outperformed the other approaches by a factor of X.
