### 2021-03-09 openmainframeproject/ade#90 @mdpiper
JOSS review: Functionality documentation
This issue is raised in reference to JOSS submission [3052](https://github.com/openjournals/joss-reviews/issues/3052).

Overall, the documentation is very helpful. However, the scoring measures used, including:

* Interval Anomaly score
* Bernoulli Score
* Best-of-two score
* Log normal score
* Poisson score
* rarity and severity scores
* critical and unique words for each message

are mentioned in the paper and in the documentation, but they aren't defined in either. Also, how are these scores combined to create the final score? This information should be in either the paper or the documentation.
