### 2021-03-17 #3052 @ayush-1506
> Where does the output of this example go?

@mdpiper The outputs generally go in `$ADE_HOME/baseline/output`. You can find the generated xml files there. This is indicated by the `BASELINE_OUTPUT_DIR` variable : [here](https://github.com/openmainframeproject/ade/blob/master/ade-assembly/src/main/test/analysis_comp_test.sh#L43).
