### 2021-03-17 openmainframeproject/ade#90 @ayush-1506
Thanks, will add this too. (I didn't want to go deep into the metrics earlier but this seems like something worth adding)
