### 2021-03-26 openmainframeproject/ade#85 @ayush-1506
Thanks for the review, @arcuri82 . Will make the changes you pointed out. I've also decided to trim the paper and change the style a bit, will update once that's done.
